pyproject.toml
[project]
name = "gen-ai-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.63",
    "langchain-groq>=0.3.2",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
